+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Stack 1-3 changes rather than only doing 1 mutation.
 - Simplifier improvements: bugfixes, faster fixed-point search.
 - Fixed issue treating failures as executions (printing . & double counting)
+- New simplifier constraint "contains output"
 
 * 1.0.0-beta.8